task1/app/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.10
gunicorn==20.1.0
requests==2.31.0
